1st commit
"75 Days Challenge" is a to-do list web application built using PHP and MySQL. It allows users to manually create an account and track their progress for 75 days. Once completed, a success banner with the user's name is displayed, which can be downloaded as a milestone achievement. The project is designed with a responsive and professional UI, making it accessible on both mobile and desktop devices

website link : http://75dayshardchallenge.lovestoblog.com